Extracts logger to external implementation allowing for overridden behavior
