+++ src/logger/ConsoleLogger.js
@@ -1,0 +1,10 @@
+
+export default class ConsoleLogger {
+
+  warn(...messages) {
+    messages.forEach(msg => {
+      console.warn(`Pubst WARNING: ${msg}`);
+    });
+  }
+
+}
