+++ src/logger/SilentLogger.js
@@ -1,0 +1,4 @@
+
+export default class SilentLogger {
+  warn() {}
+}
